Here are my 2 forks of node-sonos-http-api and node-sonos-discovery. To make it work, install the original node-sonos-http-api and then copy paste the files from these repo's over it to add additional features and fixes.
Many of the added work comes from pull requests and issues from the original project and some own work.

This is a dependancy of the node-sonos-http-api, with the following changes:

- Loudness support added
- Balance support added
- Volume is set via channel instead of master to support balance
- Set status led support added
- Error 500 fix when Spotify is connected
- Surround support added, untested

The following API's are added:

**addtoqueue**
Add track to the current queue of specified player. You can obtain track uri from /favorites/detailed call. When using Sonos music library (CIFS) and know the path of a song you can construct the uri by using encodeURIComponent. The request will accept:

*   uri (required,
*   enqueueAsNext (boolean, required)
*   desiredFirstTrackNumberEnqueued (optional)

Examples:

- http://localhost:5005/myroom/addtoqueue/x-file-cifs%3A%2F%2F10.10.1.77%2Fpublic%2FJames%2F02.%20From%20Russia%20With%20Love%2F07%20-%20007.mp3/true/1
- http://localhost:5005/myroomroom/addtoqueue/x-sonos-spotify:spotify%3atrack%3a4BggEwLhGfrbrl7JBhC8EC?sid=9&flags=8224&sn=2/true/1

**loudness**
Can only be invoked per player, not on a grouped or zone. You can call the players individually even if they are in a group or zone.

- https://github.com/jishi/node-sonos-http-api/pull/588/files
- https://sonos.barrydegraaff.tk/zolder/loudness/on
- https://sonos.barrydegraaff.tk/zolder/loudness/off

**balance**
Can only be invoked per player, not on a grouped or zone. You can call the players individually even if they are in a group or zone.  You can use balance to create a stereo pair of speakers of a different type by setting balance and then group them.

- https://github.com/jishi/node-sonos-http-api/pull/509/files
- https://github.com/jishi/node-sonos-discovery/pull/109/files
- http://localhost:5005/{room name}/balance/{integer between -100 and 100}
- https://sonos.barrydegraaff.tk/zolder/balance/0 balance to the left
- https://sonos.barrydegraaff.tk/zolder/balance/50 balance to normal
- https://sonos.barrydegraaff.tk/zolder/balance/100 balance to right

**led**
Can only be invoked per player, not on a grouped or zone. You can call the players individually even if they are in a group or zone.
- https://sonos.barrydegraaff.tk/Woonkamer/led/on
- https://sonos.barrydegraaff.tk/Woonkamer/led/off
- https://github.com/jishi/node-sonos-discovery/pull/89